get rid of files with a single context inside
