--- frontend/src/components/ui/blockEditor/directoryViewer/ChatHistoryContext.jsx
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const ChatHistoryContext = createContext();
