--- frontend/src/components/ui/blockEditor/directoryViewer/ConfirmModalContext.jsx
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const ConfirmModalContext = createContext();
