--- frontend/src/components/ui/blockEditor/directoryViewer/FileBufferContext.jsx
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const FileBufferContext = createContext(null);
