--- frontend/src/components/ui/blockEditor/directoryViewer/FileHandleContext.jsx
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const FileHandleContext = createContext();
