--- frontend/src/components/ui/blockEditor/directoryViewer/SelectedPromptContext.jsx
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const SelectedPromptContext = createContext();
